Обновлены сервисы SoundRequest для поддержки работы с плавающей точкой в громкости и улучшения структуры данных.
- Изменены типы громкости с int на float в SoundRequestController, MainPlayer, SoundRequestManager и StateManager для более точного управления.
- Обновлены сообщения об ошибках и документация для отражения нового диапазона громкости (0.0 - 100.0).
- Удалены устаревшие навигационные свойства из модели TwitchUser для упрощения структуры данных.

+++ Services/SoundRequest/Entities/PlayerState.cs
@@ -30,7 +30,10 @@ public class PlayerState
     /// </summary>
     public bool IsMuted { get; set; }
 
-    public int Volume { get; set; } = 100;
+    /// <summary>
+    /// Громкость (0.0 - 100.0)
+    /// </summary>
+    public float Volume { get; set; } = 100f;
 
     /// <summary>
     /// Twitch ID пользователя, заказавшего текущий трек
